+++ src/main/kotlin/icu/windea/pls/lang/diff/actions/CompareDefinitionsAction.kt
@@ -53,7 +53,7 @@ import javax.swing.Icon
  * 对当前定义与包括其只读副本在内的拥有相同名称和主要类型的定义进行差异比较。
  *
  * - 忽略直接位于游戏或模组的根目录下的文件。
- * - 按照覆盖方式进行排序。
+ * - 按照覆盖策略进行排序。
  */
 class CompareDefinitionsAction : ParadoxShowDiffAction() {
     private fun findSourceFile(e: AnActionEvent): VirtualFile? {