+++ src/main/kotlin/icu/windea/pls/lang/overrides/ParadoxOverrideService.kt
@@ -29,8 +29,8 @@ import icu.windea.pls.script.psi.ParadoxScriptScriptedVariable
 
 object ParadoxOverrideService {
     /**
-     * 得到目标（文件、全局封装变量、定义、定值变量、本地化等）使用的覆盖方式。
-     * 如果返回 `null`，则表示不适用覆盖方式。
+     * 得到目标（文件、全局封装变量、定义、定值变量、本地化等）使用的覆盖策略。
+     * 如果返回 `null`，则表示不适用覆盖策略。
      */
     fun getOverrideStrategy(target: Any): ParadoxOverrideStrategy? {
         val gameType by lazy { selectGameType(target) }
@@ -41,8 +41,8 @@ object ParadoxOverrideService {
     }
 
     /**
-     * 从查询参数得到目标（文件、全局封装变量、定义、定值变量、本地化等）使用的覆盖方式。
-     * 如果返回 `null`，则表示不适用覆盖方式。
+     * 从查询参数得到目标（文件、全局封装变量、定义、定值变量、本地化等）使用的覆盖策略。
+     * 如果返回 `null`，则表示不适用覆盖策略。
      */
     fun getOverrideStrategy(searchParameters: ParadoxSearchParameters<*>): ParadoxOverrideStrategy? {
         val gameType = searchParameters.selector.gameType
@@ -65,7 +65,7 @@ object ParadoxOverrideService {
 
     /**
      * 检查是否存在对文件的重载。
-     * 如果返回 `null`，则表示使用的覆盖方式为 `ORDERED`，或者不存在重载。
+     * 如果返回 `null`，则表示使用的覆盖策略为 `ORDERED`，或者不存在重载。
      */
     fun getOverrideResultForFile(file: PsiFile): ParadoxOverrideResult<PsiFile>? {
         val overrideStrategy = getOverrideStrategy(file) ?: return null
@@ -82,7 +82,7 @@ object ParadoxOverrideService {
 
     /**
      * 检查是否存在对（全局）封装变量的重载。
-     * 如果返回 `null`，则表示使用的覆盖方式为 `ORDERED`，或者不存在重载。
+     * 如果返回 `null`，则表示使用的覆盖策略为 `ORDERED`，或者不存在重载。
      */
     fun getOverrideResultForGlobalScriptedVariable(element: ParadoxScriptScriptedVariable, file: PsiFile): ParadoxOverrideResult<ParadoxScriptScriptedVariable>? {
         val name = element.name
@@ -99,7 +99,7 @@ object ParadoxOverrideService {
 
     /**
      * 检查是否存在对定义的重载。
-     * 如果返回 `null`，则表示使用的覆盖方式为 `ORDERED`，或者不存在重载。
+     * 如果返回 `null`，则表示使用的覆盖策略为 `ORDERED`，或者不存在重载。
      */
     fun getOverrideResultForDefinition(element: ParadoxScriptProperty, file: PsiFile): ParadoxOverrideResult<ParadoxScriptProperty>? {
         val definitionInfo = element.definitionInfo ?: return null
@@ -118,7 +118,7 @@ object ParadoxOverrideService {
 
     /**
      * 检查是否存在对定值变量的重载。
-     * 如果返回 `null`，则表示使用的覆盖方式为 `ORDERED`，或者不存在重载。
+     * 如果返回 `null`，则表示使用的覆盖策略为 `ORDERED`，或者不存在重载。
      */
     fun getOverrideResultForDefineVariable(element: ParadoxScriptProperty, file: PsiFile): ParadoxOverrideResult<ParadoxScriptProperty>? {
         val defineVariableInfo = element.defineVariableInfo ?: return null